+++ src/components/domains/team/TeamMemberSettingCard/index.tsx
@@ -1,0 +1,1 @@
+export { TeamMemberSettingCard } from './TeamMemberSettingCard';
